requirements.txt
discord.py
yt-dlp
newsapi-python
python-dotenv
requests